Use timestamp as provided by log record when formatting JSON
The log record provides a timestamp which is converted to text by a `DateTimeFormatter` instance.
Instead of always truncating the timestamp to millisecond precision, the original timestamp is now
given to the `DateTimeFormatter` so that it can format it however it wants. The default formatter is
still `DateTimeFormatter.ISO_OFFSET_DATE_TIME`. For timestamps with sub-millisecond precision it will
also include those fractional seconds in the formatted text representation (up to nanosecond precision).